Fix InstallerStateTest.RemoveOldVersionDirs by fixing the comparison in alternate_version_generator.cc to use max(DWORD) rather than max(size_t) since CreateFileMapping takes a DWORD.
BUG=236376
TEST=installer_util_unittests.exe --gtest_filter=InstallerStateTest.RemoveOldVersionDirs

Review URL: https://chromiumcodereview.appspot.com/14105014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@197434 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d6da48d0d797d46ad7c0fc69d9994bca02d621c8

+++ chrome/installer/test/alternate_version_generator.cc
@@ -179,7 +179,7 @@ bool MappedFile::Initialize(base::PlatformFile file) {
 
   if (base::GetPlatformFileInfo(file, &file_info)) {
     if (file_info.size <=
-        static_cast<int64>(std::numeric_limits<size_t>::max())) {
+        static_cast<int64>(std::numeric_limits<DWORD>::max())) {
       mapping_ = CreateFileMapping(file, NULL, PAGE_READWRITE, 0,
                                    static_cast<DWORD>(file_info.size), NULL);
       if (mapping_ != NULL) {
@@ -194,7 +194,7 @@ bool MappedFile::Initialize(base::PlatformFile file) {
         PLOG(DFATAL) << "CreateFileMapping failed";
       }
     } else {
-      LOG(DFATAL) << "Files larger than " << std::numeric_limits<size_t>::max()
+      LOG(DFATAL) << "Files larger than " << std::numeric_limits<DWORD>::max()
                   << " are not supported.";
     }
   } else {
